fix: add public read RLS policies and extract buttonVariants for server components
- Add migration 00037 with public SELECT RLS policies for services, reviews,
  users (doctors), time_slots, appointments, blog_posts, products, stock,
  on_duty_schedule, before_after_photos, and prescription_requests tables.
  This allows anonymous visitors to view public-facing clinic data.

- Extract buttonVariants from client-only button.tsx into a server-safe
  button-variants.ts module. Update 7 server component pages to import
  from the new module, fixing 'cannot call client function from server'
  errors on services, testimonials, doctor-profile, doctor-services,
  dentist gallery, dentist prices, and dentist services pages.
