+++ src/app/(public)/services/page.tsx
@@ -4,7 +4,7 @@ import Link from "next/link";
 import { getPublicServices } from "@/lib/data/public";
 import { defaultWebsiteConfig } from "@/lib/website-config";
 import { safeJsonLdStringify } from "@/lib/json-ld";
-import { buttonVariants } from "@/components/ui/button";
+import { buttonVariants } from "@/components/ui/button-variants";
 import { Card, CardContent, CardFooter, CardHeader, CardTitle, CardDescription } from "@/components/ui/card";
 
 export const metadata: Metadata = {
