+++ src/app/(public)/testimonials/page.tsx
@@ -3,7 +3,7 @@ import { Star, Quote, Calendar } from "lucide-react";
 import { Card, CardContent, CardHeader } from "@/components/ui/card";
 import { Avatar, AvatarFallback } from "@/components/ui/avatar";
 import { Badge } from "@/components/ui/badge";
-import { buttonVariants } from "@/components/ui/button";
+import { buttonVariants } from "@/components/ui/button-variants";
 import Link from "next/link";
 import { getPublicReviews, getPublicAverageRating, getPublicBranding } from "@/lib/data/public";
